perf: close the open hardening and usability backlog (#304)
## Summary

This branch closes the complete paginated open-issue set in
gameplay-impact order:

- removes measured P2P and spectator hot-path overhead while preserving
unknown-source diagnostics;
- replaces fragmented onboarding with one runnable first-session path
and short route maps;
- reduces the crates.io archive from 91 to 58 files and from 1,091,787
to 914,854 compressed bytes;
- audits all four authoritative Cargo workspaces in hosted
vulnerability, deny-policy, license, and freshness jobs;
- type-checks all 58 production Python modules with strict, pinned mypy
at the Python 3.10 floor.

It also raises the docs tooling floor to the latest compatible
pymdown-extensions 11.0.2.

## Measured results

- All-local N=2 P2P callgrind: 743,107,916 → 627,307,916 instructions
(-15.58%).
- Warm spectator polling: zero allocations for one- and four-host rows.
- Per-packet connection-status fan-out: zero clone allocations at
N=2/4/16.
- Published archive: 36.26% fewer files, 11.57% fewer unpacked bytes,
16.21% fewer compressed bytes.
- Strict mypy: 34 initial diagnostics across 15 files → zero issues in
58 files.
- Strict pyright comparison: 495 errors on the same scope, so one
checker is maintained.

## Verification

- Latest `main` head `2363992`: all 16 workflow groups successful.
- Strict Clippy: `tokio,json` and `hot-join,tokio,json` matrices pass.
- Nextest: 2,903 default and 3,161 hot-join tests pass.
- Rustdoc: 169 pass, 54 configured ignores.
- Python: 2,098 tests pass; Ruff and strict mypy pass.
- Package clean-room verification and all 1,953 packaged library tests
pass.
- All four Cargo locks pass RustSec and cargo-deny policy; all
compatible lock resolutions are current.
- MkDocs strict build, Markdownlint, wiki validation, actionlint, and
agent preflight pass.
- Main-thread adversarial review found and fixed the package fixture
omission, global license-exception noise, cargo-deny config resolution,
passive contributor prose, and one stale release packaging assertion; no
findings remain.

## Dependency disposition

All compatible Cargo, npm, docs, and release-tooling dependencies are
current. The only newer Cargo packages exceed the Rust 1.86 MSRV:
bincode-next 3.1.1 requires Rust 1.90, and serial_test 4.0.1 requires
Rust 1.93.1.

Closes #297
Closes #299
Closes #300
Closes #302
Closes #303

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> **Medium Risk**
> Changes core P2P/protocol event handling and crates.io publish
contents, though allocation contracts and expanded supply-chain CI
reduce regression risk.
> 
> **Overview**
> This PR tightens **runtime networking**, **consumer packaging**,
**supply-chain CI**, **Python tooling**, and **onboarding docs** in one
pass.
> 
> **Performance:** Input events now carry `peer_connect_status` as a
shared `Arc<[ConnectionStatus]>` so one packet does not clone an
N-player status vector per player. All-local `P2PSession` polls skip
endpoint work after socket accounting (unknown-source diagnostics stay).
Spectator polls reuse constructor-reserved scratch for disconnect
batches instead of allocating each idle tick. Allocation contracts and a
new `p2p_all_local` benchmark lock in zero- or bounded-allocation
behavior on warmed paths.
> 
> **crates.io:** `Cargo.toml` switches from a long `exclude` list to an
anchored **`include`** allowlist (sources, licenses, README, one wire
golden test) so published tarballs stay small and self-testable; README
links are adjusted for crates.io.
> 
> **CI / supply chain:** Security and scheduled freshness jobs run in a
**four-workspace matrix** (root, fuzz, loom, godot-emscripten) with
`cargo audit`, workspace-scoped `cargo deny`, and **lockfile-scoped
license exceptions** (NCSA/MPL test-only deps). Freshness reports are
required; version lag stays advisory. Quality CI adds **blocking strict
mypy** (Python 3.10 floor, pinned 2.3.1) plus a matching pre-commit hook
for production `scripts/`.
> 
> **Docs:** README and docs home become short route maps; new **Getting
Started** is the single runnable `SyncTestSession` path (MkDocs/wiki nav
updated). User guide quick-start defers to that guide.
> 
> **Tests:** New contracts cover CI security matrices, package payload,
Python typing gate, and docs onboarding shape.
> 
> <sup>Reviewed by [Cursor Bugbot](https://cursor.com/bugbot) for commit
579a4098729acf5121631f667076b3d623ca971d. Bugbot is set up for automated
code reviews on this repo. Configure
[here](https://www.cursor.com/dashboard/bugbot).</sup>
<!-- /CURSOR_SUMMARY -->

+++ scripts/docs/check-theme-colors.py
@@ -256,7 +256,7 @@ def relative_luminance(hex_color: str) -> float:
     def channel(value: float) -> float:
         if value <= 0.03928:
             return value / 12.92
-        return ((value + 0.055) / 1.055) ** 2.4
+        return float(((value + 0.055) / 1.055) ** 2.4)
 
     return 0.2126 * channel(red) + 0.7152 * channel(green) + 0.0722 * channel(blue)
